Rename PresetTheme annotation
