+++ app/src/main/java/com/marverenic/music/data/store/PreferenceStore.java
@@ -3,7 +3,7 @@ package com.marverenic.music.data.store;
 import android.media.audiofx.Equalizer;
 
 import com.marverenic.music.data.annotations.BaseTheme;
-import com.marverenic.music.data.annotations.PresetTheme;
+import com.marverenic.music.data.annotations.PrimaryTheme;
 import com.marverenic.music.data.annotations.StartPage;
 
 import java.util.Collection;
@@ -19,9 +19,9 @@ public interface PreferenceStore extends ReadOnlyPreferenceStore {
     void setOpenNowPlayingOnNewQueue(boolean openNowPlayingOnNewQueue);
     void setEnableNowPlayingGestures(boolean enabled);
     void setDefaultPage(@StartPage int defaultPage);
-    void setPrimaryColor(@PresetTheme int colorChoice);
+    void setPrimaryColor(@PrimaryTheme int colorChoice);
     void setBaseColor(@BaseTheme int theme);
-    void setIconColor(@PresetTheme int theme);
+    void setIconColor(@PrimaryTheme int theme);
 
     void toggleShuffle();
     void setShuffle(boolean shuffle);
