+++ app/src/main/java/com/marverenic/music/data/store/SharedPreferenceStore.java
@@ -8,7 +8,7 @@ import android.support.annotation.StringRes;
 
 import com.marverenic.music.R;
 import com.marverenic.music.data.annotations.BaseTheme;
-import com.marverenic.music.data.annotations.PresetTheme;
+import com.marverenic.music.data.annotations.PrimaryTheme;
 import com.marverenic.music.data.annotations.StartPage;
 import com.marverenic.music.player.MusicPlayer;
 
@@ -118,7 +118,7 @@ public class SharedPreferenceStore implements PreferenceStore {
     @Override
     @SuppressWarnings("WrongConstant")
     public int getPrimaryColor() {
-        return getInt(R.string.pref_key_color_primary, PresetTheme.BLUE);
+        return getInt(R.string.pref_key_color_primary, PrimaryTheme.BLUE);
     }
 
     @Override
@@ -130,7 +130,7 @@ public class SharedPreferenceStore implements PreferenceStore {
     @Override
     public int getIconColor() {
         //noinspection WrongConstant
-        return getInt(R.string.pref_key_color_icon, PresetTheme.BLUE);
+        return getInt(R.string.pref_key_color_icon, PrimaryTheme.BLUE);
     }
 
     @Override
@@ -216,7 +216,7 @@ public class SharedPreferenceStore implements PreferenceStore {
     }
 
     @Override
-    public void setPrimaryColor(@PresetTheme int colorChoice) {
+    public void setPrimaryColor(@PrimaryTheme int colorChoice) {
         putInt(R.string.pref_key_color_primary, colorChoice);
     }
 
@@ -226,7 +226,7 @@ public class SharedPreferenceStore implements PreferenceStore {
     }
 
     @Override
-    public void setIconColor(@PresetTheme int theme) {
+    public void setIconColor(@PrimaryTheme int theme) {
         putInt(R.string.pref_key_color_icon, theme);
     }
